requirements.txt
# Sarvam AI STT & TTS
sarvamai>=0.1.0
pyaudio>=0.2.14

# Redis Agent Memory Server
agent-memory-client>=0.14.0
redis>=5.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0

# Note: Requires Python 3.10+
# Agent Memory Server runs via Docker:
# docker run -p 8001:8000 -e REDIS_URL=... -e OPENAI_API_KEY=... redislabs/agent-memory-server:latest

